% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{CV.SuperLearner_tidiers}
\alias{CV.SuperLearner_tidiers}
\alias{tidy.CV.SuperLearner}
\title{Tidying method(s) for a CV.SuperLearner fit object}
\usage{
\method{tidy}{CV.SuperLearner}(x, ...)
}
\arguments{
\item{x}{object of class \code{\link[SuperLearner]{CV.SuperLearner}}}

\item{...}{Passed through to \code{\link{tidy.SuperLearner}}. Optional
arguments include \code{algorithm}, \code{includeAll}, and
\code{stringsAsFactors}. See \code{\link{tidy.SuperLearner}} for
documentation.}
}
\value{
A \code{data.frame} without rownames. Column names included depend
on the supplied value for the optional \code{\link{tidy.SuperLearner}}
argument \code{algorithm}. See \code{\link{tidy.SuperLearner}} for more
details on returned \code{data.frame}. Note that the resulting
\code{data.frame} from \code{tidy.CV.SuperLearner} will contain one
additional column, however: "fold," indicating the (outer) SuperLearner
cross-validation fold number.
}
\description{
This method tidies the details of a CV.SuperLearner fit (screening,
prediction, or both) into a summary.
}
\examples{
# based on an example in the SuperLearner package
set.seed(1)
n <- 100
p <- 10
X <- matrix(rnorm(n*p), nrow = n, ncol = p)
X <- data.frame(X)
Y <- rbinom(n, 1, plogis(.2*X[, 1] + .1*X[, 2] - .2*X[, 3] + .1*X[, 3]*X[, 4] - .2*abs(X[, 4])))

library(SuperLearner)
cvsl = CV.SuperLearner(Y, X, family = binomial(),
                       SL.library = list(c("SL.mean", "screen.FSelector.oneR"),
                                         c("SL.mean", "screen.wgtd.ttest"),
                                         c("SL.glm", "screen.FSelector.oneR"),
                                         c("SL.glm", "screen.wgtd.ttest")),
                       cvControl = list(V = 2),
                       innerCvControl = list(list(V = 2)))

tidy(cvsl)
tidy(cvsl, algorithm = "screening")
tidy(cvsl, algorithm = "both")

}
