% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{rownames_to_column.matrix}
\alias{rownames_to_column.matrix}
\title{rownames_to_column.matrix() was inspired by tibble::rownames_to_column()}
\usage{
rownames_to_column.matrix(mat, var = "rowname", stringsAsFactors = FALSE)
}
\description{
rownames_to_column.matrix() was inspired by tibble::rownames_to_column()
}
