% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fselector.R
\name{screen.FSelector.cfs}
\alias{screen.FSelector.cfs}
\title{Correlation Feature Selection (CFS) screening algorithm}
\usage{
screen.FSelector.cfs(Y, X, family, verbose = FALSE, ...)
}
\arguments{
\item{Y}{Outcome (numeric vector). See \code{\link[SuperLearner]{SuperLearner}}
for specifics.}

\item{X}{Predictor variable(s) (data.frame or matrix). See
\code{\link[SuperLearner]{SuperLearner}} for specifics.}

\item{family}{Error distribution to be used in the model:
\code{\link[stats]{gaussian}} or \code{\link[stats]{binomial}}.
Currently unused. See \code{\link[SuperLearner]{SuperLearner}}
for specifics.}

\item{verbose}{Should debugging messages be printed? Default: \code{FALSE}.}

\item{...}{Currently unused.}
}
\value{
A logical vector with length equal to \code{ncol(X)}.
}
\description{
CFS (Hall, 1999) utilizes \code{\link[FSelector]{best.first.search}} to find
columns of \code{X} correlated with \code{Y} but not with one another (i.e.,
not redundant). CFS, combined with a search algorithm, does not rank features
and therefore does not allow for specification of either the number of
features to be chosen (\code{k}) or the method by which they should be chosen
(\code{selector}).
}
\examples{
data(iris)
Y <- as.numeric(iris$Species=="setosa")
X <- iris[,-which(colnames(iris)=="Species")]
screen.FSelector.cfs(Y, X, binomial())

data(mtcars)
Y <- mtcars$mpg
X <- mtcars[,-which(colnames(mtcars)=="mpg")]
screen.FSelector.cfs(Y, X, gaussian())

# based on examples in SuperLearner package
set.seed(1)
n <- 100
p <- 20
X <- matrix(rnorm(n*p), nrow = n, ncol = p)
X <- data.frame(X)
Y <- X[, 1] + sqrt(abs(X[, 2] * X[, 3])) + X[, 2] - X[, 3] + rnorm(n)

library(SuperLearner)
sl = SuperLearner(Y, X, family = gaussian(), cvControl = list(V = 2),
                  SL.library = list(c("SL.glm", "All"),
                                    c("SL.glm.interaction", "screen.FSelector.cfs")))
sl
sl$whichScreen
}
\references{
\url{http://citeseerx.ist.psu.edu/viewdoc/summary?doi=10.1.1.37.4643}
}
