% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fselector.R
\name{screen.FSelector.consistency}
\alias{screen.FSelector.consistency}
\title{Consistency screening algorithm}
\usage{
screen.FSelector.consistency(Y, X, family, verbose = FALSE, ...)
}
\arguments{
\item{Y}{Outcome (numeric vector). See \code{\link[SuperLearner]{SuperLearner}}
for specifics.}

\item{X}{Predictor variable(s) (data.frame or matrix). See
\code{\link[SuperLearner]{SuperLearner}} for specifics.}

\item{family}{Error distribution to be used in the model:
\code{\link[stats]{gaussian}} or \code{\link[stats]{binomial}}.
Currently unused. See \code{\link[SuperLearner]{SuperLearner}}
for specifics.}

\item{verbose}{Should debugging messages be printed? Default: \code{FALSE}.}

\item{...}{Currently unused.}
}
\value{
A logical vector with length equal to \code{ncol(X)}.
}
\description{
The \code{\link[FSelector]{consistency}} algorithm utilizes
\code{\link[FSelector]{best.first.search}} to find the optimal combination of
columns of \code{X} to minimize 'inconsistency' with the outcome \code{Y}.
Implemented for \code{binomial()} family only and designed to be used with
binary or categorical \code{X}. Continuous \code{X} will be discretized by
\code{\link{FSelector}} and \code{\link[RWeka]{Discretize}} using the MDL
method (Fayyad & Irani, 1993). Search algorithms do not rank features and
therefore this algorithm does not allow for specification of either the
number of features to be chosen (\code{k}) or the method by which they
should be chosen (\code{selector}).
}
\examples{
data(iris)
Y <- as.numeric(iris$Species=="setosa")
X <- iris[,-which(colnames(iris)=="Species")]
screen.FSelector.consistency(Y, X, binomial())

# based on example in SuperLearner package
set.seed(1)
n <- 100
p <- 20
X <- matrix(rnorm(n*p), nrow = n, ncol = p)
X <- data.frame(X)
Y <- rbinom(n, 1, plogis(.2*X[, 1] + .1*X[, 2] - .2*X[, 3] + .1*X[, 3]*X[, 4] - .2*abs(X[, 4])))

library(SuperLearner)
sl = SuperLearner(Y, X, family = binomial(), cvControl = list(V = 2),
                  SL.library = list(c("SL.lm", "All"),
                                    c("SL.lm", "screen.FSelector.consistency")))
sl
sl$whichScreen
}
\references{
\url{http://hdl.handle.net/2014/35171}
}
