% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fselector.R
\name{screen.FSelector}
\alias{screen.FSelector}
\title{Screening algorithms implemented in the FSelector package}
\usage{
screen.FSelector(Y, X, family, obsWeights, id, filter = c("cfs",
  "chi.squared", "consistency", "gain.ratio", "information.gain",
  "linear.correlation", "oneR", "random.forest.importance", "rank.correlation",
  "relief", "symmetrical.uncertainty"), filter_params = NULL,
  selector = c("cutoff.biggest.diff", "cutoff.k", "cutoff.k.percent", "all"),
  k = switch(selector, cutoff.k = ceiling(0.5 * ncol(X)), cutoff.k.percent =
  0.5, NULL), verbose = FALSE, ...)
}
\arguments{
\item{Y}{Outcome (numeric vector). See \code{\link[SuperLearner]{SuperLearner}}
for specifics.}

\item{X}{Predictor variable(s) (data.frame or matrix). See
\code{\link[SuperLearner]{SuperLearner}} for specifics.}

\item{family}{Error distribution to be used in the model:
\code{\link[stats]{gaussian}} or \code{\link[stats]{binomial}}.
Currently unused. See \code{\link[SuperLearner]{SuperLearner}}
for specifics.}

\item{obsWeights}{Optional numeric vector of observation weights. Currently
unused.}

\item{id}{Cluster identification variable. Currently unused.}

\item{filter}{A string corresponding to a feature ranking or selecting function
implemented in the FSelector package. One of: \code{\link[FSelector]{cfs}},
\code{\link[FSelector]{chi.squared}}, \code{\link[FSelector]{consistency}},
\code{\link[FSelector]{gain.ratio}}, \code{\link[FSelector]{information.gain}},
\code{\link[FSelector]{linear.correlation}}, \code{\link[FSelector]{oneR}},
\code{\link[FSelector]{random.forest.importance}},
\code{\link[FSelector]{rank.correlation}}, \code{\link[FSelector]{relief}},
\code{\link[FSelector]{symmetrical.uncertainty}}. Default: \code{"cfs"}.
Note that "filter" is a misnomer in the case of embedded feature selection
methods such as \code{\link[FSelector]{random.forest.importance}}.}

\item{filter_params}{A named list of tuning parameter arguments specific to
the chosen \code{filter}. Default of \code{NULL} should be used when either
1) the chosen \code{filter} does not utilize tuning parameter(s) or
2) the default values should be retained.}

\item{selector}{A string corresponding to a subset selecting function
implemented in the FSelector package. One of:
\code{\link[FSelector]{cutoff.biggest.diff}},
\code{\link[FSelector]{cutoff.k}}, \code{\link[FSelector]{cutoff.k.percent}},
or \code{"all"}. Note that \code{"all"} is a not a function but indicates
pass-thru should be performed in the case of a \code{filter} which selects
rather than ranks features. Default: \code{"cutoff.biggest.diff"}.}

\item{k}{Passed through to the \code{selector} in the case where \code{selector} is
\code{\link[FSelector]{cutoff.k}} or \code{\link[FSelector]{cutoff.k.percent}}.
Otherwise, should remain NULL (the default). For \code{\link[FSelector]{cutoff.k}},
this is an integer indicating the number of features to keep from \code{X}.
For \code{\link[FSelector]{cutoff.k.percent}}, this is instead the proportion
of features to keep.}

\item{verbose}{Should debugging messages be printed? Default: \code{FALSE}.}

\item{...}{Currently unused.}
}
\value{
A logical vector with length equal to \code{ncol(X)}.
}
\description{
A SuperLearner-compatible interface to functions in the
\code{\link[FSelector]{FSelector}} package.
}
