% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{screen.wgtd.corRank}
\alias{screen.wgtd.corRank}
\title{Weighted correlation screening algorithm}
\usage{
screen.wgtd.corRank(Y, X, family, obsWeights, id, method = "pearson", k = 2,
  ...)
}
\arguments{
\item{Y}{Outcome (numeric vector). See \code{\link[SuperLearner]{SuperLearner}}
for specifics.}

\item{X}{Predictor variable(s) (data.frame or matrix). See
\code{\link[SuperLearner]{SuperLearner}} for specifics.}

\item{family}{Error distribution to be used in the model:
\code{\link[stats]{gaussian}} or \code{\link[stats]{binomial}}.
Currently unused. See \code{\link[SuperLearner]{SuperLearner}}
for specifics.}

\item{obsWeights}{Optional numeric vector of observation weights. See
\code{\link[SuperLearner]{SuperLearner}} for specifics.}

\item{id}{Cluster identification variable. Currently unused.}

\item{method}{Which correlation coefficient to compute. Currently accepts
\code{"pearson"} or \code{"spearman"}.}

\item{k}{Minimum number of features to select.}

\item{...}{Currently unused.}
}
\value{
A logical vector with length equal to \code{ncol(X)}
}
\description{
Performs feature selection according to the ranking of weighted
correlation coefficient estimates. Implemented via
\code{\link[wCorr]{weightedCorr}}.
}
\examples{
# based on examples in SuperLearner package
set.seed(1)
n <- 100
p <- 20
X <- matrix(rnorm(n*p), nrow = n, ncol = p)
X <- data.frame(X)
Y <- X[, 1] + sqrt(abs(X[, 2] * X[, 3])) + X[, 2] - X[, 3] + rnorm(n)
obsWeights <- 1/runif(n)
screen.wgtd.corRank(Y, X, gaussian(), obsWeights, seq(n), k = 3)

screen.wgtd.corRank3 <- function(..., k = 3){
    screen.wgtd.corRank(..., k = k)
}

library(SuperLearner)
sl = SuperLearner(Y, X, family = gaussian(), cvControl = list(V = 2),
                  obsWeights = obsWeights,
                  SL.library = list(c("SL.glm", "All"),
                                    c("SL.glm.interaction", "screen.wgtd.corRank3")))
sl
sl$whichScreen
}
