% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twosamptests.R
\name{screen.wgtd.ttest}
\alias{screen.wgtd.ttest}
\title{Weighted t-test screening algorithm}
\usage{
screen.wgtd.ttest(
  Y,
  X,
  family,
  obsWeights,
  id,
  selector = c("cutoff.k", "cutoff.k.percent"),
  k = switch(selector, cutoff.k = ceiling(0.5 * ncol(X)), cutoff.k.percent = 0.5, NULL),
  minP = NULL,
  ...
)
}
\arguments{
\item{Y}{Outcome (numeric vector). See \code{\link[SuperLearner]{SuperLearner}}
for specifics.}

\item{X}{Predictor variable(s) (data.frame or matrix). See
\code{\link[SuperLearner]{SuperLearner}} for specifics.}

\item{family}{Error distribution to be used in the model:
\code{\link[stats]{gaussian}} or \code{\link[stats]{binomial}}.
See \code{\link[SuperLearner]{SuperLearner}} for specifics.}

\item{obsWeights}{Optional numeric vector of observation weights. See
\code{\link[SuperLearner]{SuperLearner}} for specifics.}

\item{id}{Cluster identification variable. Currently unused.}

\item{selector}{A string corresponding to a subset selecting function
implemented in the FSelector package. One of:
\code{\link[FSelector]{cutoff.k}} or \code{\link[FSelector]{cutoff.k.percent}}.
Ignored if \code{minP} is non-\code{NULL} and at least \code{k} features
have P-values at or below \code{minP}. Default: \code{"cutoff.k"}.}

\item{k}{Numeric. Minimum number or proportion of features to select.
Passed through to the \code{selector}. For \code{\link[FSelector]{cutoff.k}},
this is an integer indicating the number of features to keep from \code{X}.
For \code{\link[FSelector]{cutoff.k.percent}}, this is instead the proportion
of features to keep. Ignored if \code{minP} is non-\code{NULL} and at least
\code{k} features have P-values at or below \code{minP}.}

\item{minP}{Numeric. To pass the screen, resulting P-values must not exceed this
number. Ignored if \code{NULL} (default) or if fewer than \code{k} features
have P-values at or below this value.}

\item{...}{Passed to \code{\link[weights]{wtd.t.test}}. These arguments
control bootstrapping of P-values and standard errors as well as forced
scaling of weights.}
}
\value{
A logical vector with length equal to \code{ncol(X)}
}
\description{
Performs feature selection according to the ranking of t statistics
or P-values returned from weighted t-tests. Implemented via
\code{\link[weights]{wtd.t.test}}.
}
\examples{
# based on example in SuperLearner package
set.seed(1)
n <- 100
p <- 20
X <- matrix(rnorm(n*p), nrow = n, ncol = p)
X <- data.frame(X)
Y <- rbinom(n, 1, plogis(.2*X[, 1] + .1*X[, 2] - .2*X[, 3] + .1*X[, 3]*X[, 4] - .2*abs(X[, 4])))
obsWeights <- 1/runif(n)
screen.wgtd.ttest(Y, X, binomial(), obsWeights, seq(n), k = 4)

screen.wgtd.ttest4 <- function(..., k = 4){
    screen.wgtd.ttest(..., k = k)
}

library(SuperLearner)
sl = SuperLearner(Y, X, family = binomial(), cvControl = list(V = 2),
                  obsWeights = obsWeights,
                  SL.library = list(c("SL.lm", "All"),
                                    c("SL.lm", "screen.wgtd.ttest4")))
sl
sl$whichScreen
}
