% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{screen.wgtd.ttestP}
\alias{screen.wgtd.ttestP}
\title{Weighted t-test screening algorithm}
\usage{
screen.wgtd.ttestP(Y, X, family, obsWeights, id, minPvalue = 0.1,
  minscreen = 2, ...)
}
\arguments{
\item{Y}{Outcome (numeric vector). See \code{\link[SuperLearner]{SuperLearner}}
for specifics.}

\item{X}{Predictor variable(s) (data.frame or matrix). See
\code{\link[SuperLearner]{SuperLearner}} for specifics.}

\item{family}{Error distribution to be used in the model: \code{"gaussian"}
or \code{"binomial"}.  See \code{\link[SuperLearner]{SuperLearner}} for
specifics.}

\item{obsWeights}{Optional numeric vector of observation weights. See
\code{\link[SuperLearner]{SuperLearner}} for specifics.}

\item{id}{Cluster identification variable. Currently unused.}

\item{minPvalue}{To pass the screen, resulting P-values must not exceed this
number.}

\item{minscreen}{Minimum number of features to select (aka rank). Only used
if less than this number of features are selected using \code{minPvalue}.}

\item{...}{Passed to \code{\link[weights]{wtd.t.test}}. These arguments
control bootstrapping of P-values and standard errors as well as forced
scaling of weights.}
}
\value{
A logical vector with length equal to \code{ncol(X)}
}
\description{
Performs feature selection according to the ranking of P-values
returned from weighted t-tests.
}
