% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twosamptests.R
\name{screen.wgtd.ttestRank}
\alias{screen.wgtd.ttestRank}
\title{Weighted t-test screening algorithm}
\usage{
screen.wgtd.ttestRank(Y, X, family, obsWeights, id, minscreen = 2, ...)
}
\arguments{
\item{Y}{Outcome (numeric vector). See \code{\link[SuperLearner]{SuperLearner}}
for specifics.}

\item{X}{Predictor variable(s) (data.frame or matrix). See
\code{\link[SuperLearner]{SuperLearner}} for specifics.}

\item{family}{Error distribution to be used in the model:
\code{\link[stats]{gaussian}} or \code{\link[stats]{binomial}}.
See \code{\link[SuperLearner]{SuperLearner}} for specifics.}

\item{obsWeights}{Optional numeric vector of observation weights. See
\code{\link[SuperLearner]{SuperLearner}} for specifics.}

\item{id}{Cluster identification variable. Currently unused.}

\item{minscreen}{Minimum number of features to select (aka rank). Only used
if less than this number of features are selected using \code{minPvalue}.}

\item{...}{Passed to \code{\link[weights]{wtd.t.test}}. These arguments
control bootstrapping of P-values and standard errors as well as forced
scaling of weights.}
}
\value{
A logical vector with length equal to \code{ncol(X)}
}
\description{
Performs feature selection according to the ranking of t statistics
returned from weighted t-tests. Implemented via
\code{\link[weights]{wtd.t.test}}.
}
\examples{
# based on example in SuperLearner package
set.seed(1)
n <- 100
p <- 20
X <- matrix(rnorm(n*p), nrow = n, ncol = p)
X <- data.frame(X)
Y <- rbinom(n, 1, plogis(.2*X[, 1] + .1*X[, 2] - .2*X[, 3] + .1*X[, 3]*X[, 4] - .2*abs(X[, 4])))
obsWeights <- 1/runif(n)
screen.wgtd.ttestRank(Y, X, binomial(), obsWeights, seq(n), minscreen = 4)
}
