% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{SuperLearner_tidiers}
\alias{SuperLearner_tidiers}
\alias{tidy.SuperLearner}
\title{Tidying method(s) for a SuperLearner fit object}
\usage{
\method{tidy}{SuperLearner}(x, algorithm = c("prediction", "screening",
  "both"), stringsAsFactors = FALSE, ...)
}
\arguments{
\item{x}{object of class \code{\link[SuperLearner]{SuperLearner}}}

\item{algorithm}{one of "prediction" (the default), "screening", or "both"
(where "both" indicates that information on both "prediction" and
"screening" should be reported).}

\item{stringsAsFactors}{Set to \code{FALSE} by default. Experimental feature
 -- note that setting this to \code{TRUE} may produce warnings/errors
(and/or may have no effect).}

\item{...}{passed through to internal functions handling summaries of
feature selection and/or prediction models (depending on the supplied value
for \code{algorithm}). See details (below) for specific arguments accepted.}
}
\value{
A \code{data.frame} without rownames. Column names included depend
on the supplied value for \code{algorithm}.
}
\description{
This method extends \code{\link[broom]{tidy}} to tidy the results from a
\code{\link[SuperLearner]{SuperLearner}} fit (screening, prediction, or
both) into a summary.
}
\details{
This method can be used to summarize information related to the
screening algorithm(s), prediction algorithm(s), or both.
}
\section{Optional argument(s)}{


If \code{algorithm} is set to "screening" or "both", the optional argument
\code{includeAll} can be set to \code{TRUE} to include results from the
pass-thru screening algorithm, \code{"All"}. By default, \code{includeAll}
is set to \code{FALSE} and these results are excluded.
}

\section{Resulting \code{data.frame}}{


Columns in resulting \code{data.frame} depend upon selected \code{algorithm}:
\describe{
    \item{"prediction"}{
        One row per element in \code{SL.library}. Five columns:
        \describe{
            \item{"estimate"}{Coefficient estimate}
            \item{"cvRisk"}{Estimate of cross-validated risk.}
            \item{"screener"}{Screening algorithm}
            \item{"predictor"}{Prediction algorithm}
            \item{"discrete"}{Logical. Is this the discrete SuperLearner?}
        }
    }
    \item{"screening"}{Number of rows equal to the product of
        [number columns in \code{X}] and
        [number of unique screening algorithms in \code{SL.library}, not
        including \code{"All"} (by default)].
        Three columns:
        \describe{
            \item{"screener"}{Screening algorithm}
            \item{"term"}{Column of \code{X}}
            \item{"selected"}{Logical. Did the screening algorithm select
                  this column of \code{X}?}
        }
    }
    \item{"both"}{Number of rows equal to the product of
        [number columns in \code{X}] and
        [number of elements in \code{SL.library}, not including any elements
        where the screening algorithm was set to \code{"All"} (by default)].
        Seven columns constituting the union of the columns returned
        for "prediction" and "screening" (described above).
    }
}
}

\examples{
# based on an example in the SuperLearner package
set.seed(1)
n <- 100
p <- 10
X <- matrix(rnorm(n*p), nrow = n, ncol = p)
X <- data.frame(X)
Y <- rbinom(n, 1, plogis(.2*X[, 1] + .1*X[, 2] - .2*X[, 3] + .1*X[, 3]*X[, 4] - .2*abs(X[, 4])))

library(SuperLearner)
sl = SuperLearner(Y, X, family = binomial(), cvControl = list(V = 2),
                  SL.library = list(c("SL.mean", "screen.wgtd.corP"),
                                    c("SL.mean", "screen.wgtd.ttest"),
                                    c("SL.glm", "screen.wgtd.corP"),
                                    c("SL.glm", "screen.wgtd.ttest")))

library(broom)
tidy(sl)
tidy(sl, algorithm = "screening")
tidy(sl, algorithm = "both")

}
\seealso{
\code{\link{tidy.CV.SuperLearner}}
}
