% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twosamptests.R
\name{screen.wgtd.ttestP}
\alias{screen.wgtd.ttestP}
\title{Weighted t-test screening algorithm}
\usage{
screen.wgtd.ttestP(..., selector = "cutoff.k", k = 1, minP = 0.1)
}
\arguments{
\item{...}{Passed to \code{\link[weights]{wtd.t.test}}. These arguments
control bootstrapping of P-values and standard errors as well as forced
scaling of weights.}

\item{selector}{A string corresponding to a subset selecting function
implemented in the FSelector package. One of:
\code{\link[FSelector]{cutoff.k}} or \code{\link[FSelector]{cutoff.k.percent}}.
Ignored if \code{minP} is non-\code{NULL} and at least \code{k} features
have P-values at or below \code{minP}. Default: \code{"cutoff.k"}.}

\item{k}{Numeric. Minimum number or proportion of features to select.
Passed through to the \code{selector}. For \code{\link[FSelector]{cutoff.k}},
this is an integer indicating the number of features to keep from \code{X}.
For \code{\link[FSelector]{cutoff.k.percent}}, this is instead the proportion
of features to keep. Default: \code{1}. Ignored if at least \code{k} features
have P-values at or below \code{minP}.}

\item{minP}{Numeric. To pass the screen, resulting P-values must not exceed
this number. Default is \code{0.1}. Ignored if fewer than \code{k} features
have P-values at or below this value.}
}
\value{
A logical vector with length equal to \code{ncol(X)}
}
\description{
Convenience wrapper to perform feature selection according to the ranking of
P-values returned from weighted t-tests. Implemented via
\code{\link[weights]{wtd.t.test}}.
}
\examples{
# based on example in SuperLearner package
set.seed(1)
n <- 100
p <- 20
X <- matrix(rnorm(n*p), nrow = n, ncol = p)
X <- data.frame(X)
Y <- rbinom(n, 1, plogis(.2*X[, 1] + .1*X[, 2] - .2*X[, 3] + .1*X[, 3]*X[, 4] - .2*abs(X[, 4])))
obsWeights <- 1/runif(n)
screen.wgtd.ttestP(Y, X, binomial(), obsWeights, seq(n), minP = 0.05)

screen.wgtd.ttestP05 <- function(..., minP = 0.05){
    screen.wgtd.ttestP(..., minP = minP)
}

library(SuperLearner)
sl = SuperLearner(Y, X, family = binomial(), cvControl = list(V = 2),
                  obsWeights = obsWeights,
                  SL.library = list(c("SL.lm", "All"),
                                    c("SL.lm", "screen.wgtd.ttestP05")))
sl
sl$whichScreen
}
